+++ src/lib/services/evm/index.ts
@@ -5,7 +5,7 @@ import {
   useCurrentChain,
   useLcdEndpoint,
 } from "lib/app-provider";
-import type { HexAddr20 } from "lib/types";
+import type { Addr20, HexAddr20 } from "lib/types";
 import { isHexWalletAddress } from "lib/utils";
 
 import {
@@ -29,7 +29,7 @@ export const useEvmParams = () => {
   );
 };
 
-export const useEvmCodesByAddress = (address: HexAddr20) => {
+export const useEvmCodesByAddress = (address: Addr20, enabled = true) => {
   const lcdEndpoint = useLcdEndpoint();
 
   return useQuery(
@@ -38,7 +38,7 @@ export const useEvmCodesByAddress = (address: HexAddr20) => {
     {
       retry: 1,
       refetchOnWindowFocus: false,
-      enabled: address && isHexWalletAddress(address),
+      enabled: enabled && address && isHexWalletAddress(address),
     }
   );
 };